Finalized checkout service and controller

+++ sephora-backend/Application/Services/Implementations/CartService.cs
@@ -26,7 +26,7 @@ public class CartService(
         );
     }
 
-    public async Task<CartDto?> GetById(int id)
+    public async Task<CartDto?> GetById(long id)
     {
         CartItem? entry = await cartRepository.GetById(id);
         return entry is null ? null : mapper.Map<CartDto>(entry);
@@ -43,7 +43,7 @@ public class CartService(
         await cartRepository.Save();
     }
 
-    public async Task Delete(int id)
+    public async Task Delete(long id)
     {
         await cartRepository.Delete(id);
         await cartRepository.Save();
